requirements.txt
asgiref==3.4.1
backcall==0.2.0
black==21.10b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.9
django-filter==21.1
django-nested-inline==0.4.4
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipython==7.29.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.2
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.8.4
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tomli==1.2.2
traitlets==5.1.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
